sstable: split single-level iter out of reader.go
